requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
pydantic>=2.0.0
python-json-logger>=2.0.0
tqdm>=4.65.0